requirements.txt
numpy==1.25.0
opencv-python==4.7.0.72
pandas==2.0.2
Pillow==9.5.0
PyAutoGUI==0.9.54
pytesseract==0.3.10
pywin32==306
pywinauto==0.6.8
